Refactor wasm test fixtures to include only necessary code and files.
