+++ tests/wasm/test_fixtures/README.md
@@ -1,5 +1,7 @@
 ## Rebuilding the fixtures
 
 ```
-> crust contracts build --optimized ./simple_state
+> crust contracts build --optimized ./simple-state
+
+> crust contracts build --optimized ./bank-send
 ```
