+++ tests/wasm/test_fixtures/simple-state/Cargo.toml
