+++ tests/wasm/test_fixtures/simple-state/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod contract;
