--- tests/wasm/test_fixtures/simple_state/src/lib.rs
@@ -1,8 +1,0 @@
-pub mod contract;
-mod error;
-pub mod helpers;
-pub mod integration_tests;
-pub mod msg;
-pub mod state;
-
-pub use crate::error::ContractError;
